+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - fj-doc-sample, fj-doc-mod-openpdf version set to 1.3.0
+- fj-doc-sample, setup fop-config fod pdf/a doc handler as inline
 
 ### Fixed